Cleaned up plugins, fixed indentation issue.
this was caused by deprication of mxw/vim-jsx
where formatting wasn't happening.

removed or otherwise diabled unused plugins, and their corresponding
settings.

Quicktip: ':map' as well as ':verbose set <somecommand>?' are amazing
in this case I used ':verbose set ai?' to find out that a function from
the JSX formatter was being called. After that it was trivial to fix by
dropping in a supported replacement.